Activating listview to go to other activies
I have added to method which facilitates that different part of the listview can be pressed and lead of to other activities. The strings
"final String [] values = new String[]{"den første", "den anden"}" are the items which will show up on the listview - they should be changed when Rasmus and Babis get on with their parts, as to reflect the cards. Also the "nameOfListView" needs to be named correctly to fit the listview.  And least but not last the activity1.class and activity2.class need to be replaced with the actual activities, which we want to launch.